+++ BepInEx.SplashScreen.GUI/Program.cs
@@ -16,7 +16,7 @@ namespace BepInEx.SplashScreen
     {
         private static SplashScreen _mainForm;
 
-        private static readonly System.Timers.Timer _AliveTimer = new System.Timers.Timer(30000); //30 seconds
+        private static readonly System.Timers.Timer _AliveTimer = new System.Timers.Timer(45000); //45 seconds
 
         /// <summary>
         /// The main entry point for the application.
@@ -34,7 +34,7 @@ namespace BepInEx.SplashScreen
                     if (MessageBox.Show("This is a splash screen that shows loading progress when a game patched with BepInEx is loading. It is automatically started and then updated by \"BepInEx.SplashScreen.Patcher.dll\" and can't be opened manually.\n\n" +
                                         "If you can't see a splash screen when the game is starting:\n" +
                                         "1 - Make sure that \"LoadingScreen.GUI.exe\" and \"BepInEx.SplashScreen.Patcher.dll\" are both present inside the \"BepInEx\\patchers\" folder.\n" +
-                                        "2 - Check if the splash screen isn't disabled in \"BepInEx\\config\\BepInEx.cfg\".\n" +
+                                        "2 - Check if the splash screen isn't disabled in \"BepInEx\\config\\Bertogim.LoadingScreen.cfg\".\n" +
                                         "3 - Update BepInEx5 to latest version and make sure that it is running.\n" +
                                         "4 - If the splash screen still does not appear, check the game log for any errors or exceptions. You can report issues on GitHub.\n\n" +
                                         "Do you want to open the GitHub repository page of BepInEx.SplashScreen?",
@@ -200,7 +200,7 @@ namespace BepInEx.SplashScreen
                 if (Directory.Exists(bepInExPath))
                 {
                     string configDir = Path.Combine(bepInExPath, "config");
-                    string configPath = Path.Combine(configDir, "BepInEx.cfg");
+                    string configPath = Path.Combine(configDir, "Bertogim.LoadingScreen.cfg");
 
                     if (File.Exists(configPath))
                     {
@@ -211,7 +211,7 @@ namespace BepInEx.SplashScreen
                         {
                             string trimmedLine = line.Trim();
 
-                            // Find the [SplashScreen] section
+                            // Find the [LoadingScreen] section
                             if (trimmedLine.StartsWith($"[{section}]"))
                             {
                                 inTargetSection = true;
@@ -246,8 +246,8 @@ namespace BepInEx.SplashScreen
             return defaultValue;
         }
 
-        public static string SplashScreenWindowType => GetBepInExConfigValue("SplashScreen", "WindowType", "FakeGame");
-        public static int SplashScreenExtraWaitTime => int.Parse(GetBepInExConfigValue("SplashScreen", "ExtraWaitTime", "5"));
+        public static string SplashScreenWindowType => GetBepInExConfigValue("LoadingScreen", "WindowType", "FakeGame");
+        public static int SplashScreenExtraWaitTime => int.Parse(GetBepInExConfigValue("LoadingScreen", "ExtraWaitTime", "5"));
 
         private static void ProcessInputMessage(string message)
         {